[mesh-forwarder] MTD code size optimizations
