+++ etc/visual-studio/libopenthread_k.vcxproj.filters
@@ -219,6 +219,12 @@
     <ClCompile Include="..\..\src\core\thread\mesh_forwarder.cpp">
       <Filter>Source Files\thread</Filter>
     </ClCompile>
+    <ClCompile Include="..\..\src\core\thread\mesh_forwarder_ftd.cpp">
+      <Filter>Source Files\thread</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\src\core\thread\mesh_forwarder_mtd.cpp">
+      <Filter>Source Files\thread</Filter>
+    </ClCompile>
     <ClCompile Include="..\..\src\core\thread\mle.cpp">
       <Filter>Source Files\thread</Filter>
     </ClCompile>
