+++ src/core/thread/mesh_forwarder_mtd.cpp
@@ -1,0 +1,92 @@
+/*
+ *  Copyright (c) 2018, The OpenThread Authors.
+ *  All rights reserved.
+ *
+ *  Redistribution and use in source and binary forms, with or without
+ *  modification, are permitted provided that the following conditions are met:
+ *  1. Redistributions of source code must retain the above copyright
+ *     notice, this list of conditions and the following disclaimer.
+ *  2. Redistributions in binary form must reproduce the above copyright
+ *     notice, this list of conditions and the following disclaimer in the
+ *     documentation and/or other materials provided with the distribution.
+ *  3. Neither the name of the copyright holder nor the
+ *     names of its contributors may be used to endorse or promote products
+ *     derived from this software without specific prior written permission.
+ *
+ *  THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+ *  AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+ *  IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+ *  ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
+ *  LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
+ *  CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
+ *  SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
+ *  INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
+ *  CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+ *  ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+ *  POSSIBILITY OF SUCH DAMAGE.
+ */
+
+/**
+ * @file
+ *   This file implements MTD-specific mesh forwarding of IPv6/6LoWPAN messages.
+ */
+
+#if OPENTHREAD_MTD
+
+#define WPP_NAME "mesh_forwarder_mtd.tmh"
+
+#include "mesh_forwarder.hpp"
+
+namespace ot {
+
+otError MeshForwarder::SendMessage(Message &aMessage)
+{
+    otError error;
+
+    aMessage.SetDirectTransmission();
+    aMessage.SetOffset(0);
+    aMessage.SetDatagramTag(0);
+
+    SuccessOrExit(error = mSendQueue.Enqueue(aMessage));
+    mScheduleTransmissionTask.Post();
+
+exit:
+    return error;
+}
+
+otError MeshForwarder::EvictIndirectMessage(void)
+{
+    return OT_ERROR_NOT_FOUND;
+}
+
+otError MeshForwarder::GetIndirectTransmission(void)
+{
+    return OT_ERROR_NOT_FOUND;
+}
+
+otError MeshForwarder::RemoveMessageFromSleepyChild(Message &aMessage, Child &aChild)
+{
+    OT_UNUSED_VARIABLE(aMessage);
+    OT_UNUSED_VARIABLE(aChild);
+    return OT_ERROR_NOT_FOUND;
+}
+
+void MeshForwarder::HandleSentFrameToChild(const Mac::Frame &aFrame, otError aError, const Mac::Address &aMacDest)
+{
+    OT_UNUSED_VARIABLE(aFrame);
+    OT_UNUSED_VARIABLE(aError);
+    OT_UNUSED_VARIABLE(aMacDest);
+}
+
+void MeshForwarder::HandleMesh(uint8_t *aFrame, uint8_t aFrameLength, const Mac::Address &aMacSource,
+                               const otThreadLinkInfo &aLinkInfo)
+{
+    OT_UNUSED_VARIABLE(aFrame);
+    OT_UNUSED_VARIABLE(aFrameLength);
+    OT_UNUSED_VARIABLE(aMacSource);
+    OT_UNUSED_VARIABLE(aLinkInfo);
+}
+
+}  // namespace ot
+
+#endif  // OPENTHREAD_MTD
